mypy lint
