+++ ksim_kbot/misc_tasks/data/README.md
@@ -1,0 +1,12 @@
+# Motion reference examples
+
+# cheer-hands-waving-dance
+To map the actorcore motion to K-Bot update offsets (after converting from .fbx to .bvh):
+
+- CC_Base_Pelvis: OFFSET -0.000000 -15.000001 30.000069
+- CC_Base_L_Thigh: OFFSET 9.679232 0.301105 -2.207489
+- CC_Base_L_Foot: OFFSET -0.952431 0.632389 -49.810875
+- CC_Base_R_Thigh: OFFSET -9.679712 0.301104 -2.207550
+- CC_Base_R_Calf: OFFSET 0.950368 0.643215 -49.811081
+- CC_Base_L_UpperarmTwist01: 0.000006 0.000000 -0.000092
+- CC_Base_R_UpperarmTwist01: 0.000004 -0.000002 0.000061
